Sarsa is sensitive to weight initialisation
Badly initialised, they cause vanishing/ exploding gradients ( will check what was it exactly). Eve with the nice reward scheme this happens. I just attach pic of the error. So for salsa to avoid gradients issues we need to either change weight initialisation or change optimiser. Adagrad is not sensitive to weight initialisation.